% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes_identify.R
\name{axes_identify}
\alias{axes_identify}
\title{axes_identify
Identifies plot axes location}
\usage{
axes_identify(fig.hsl, bg_lightness = 0.1)
}
\arguments{
\item{fig.hsl}{array of pixels in HSL format (hue/saturation/lightness)}

\item{bg_lightness}{a lightness threshold value between 0 and 1; every pixel with lightness > bg_lightness is considered background and removed (default: 0.1).}
}
\value{
a list w/ 2 objects, fig.hsl: the array of pixels cropped to the axes, axes: a list of xaxis, yaxis: vectors of integers to index plot by columns and rows respectively.
}
\description{
axes_identify
Identifies plot axes location
}
\examples{
# axes_identify(fig.hsl = figure, bg_lightness = 0.1)
}
