% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detect.R
\name{eventdetect}
\alias{eventdetect}
\title{eventdetect
detects potential events}
\usage{
eventdetect(res_list)
}
\arguments{
\item{res_list}{: output from overlap_detect}
}
\value{
res_list2: a dataframe with potential x, and y values
}
\description{
eventdetect
detects potential events
}
\examples{
# eventdetect(overlap_detectout)
}
