% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_clean.R
\name{fig_clean}
\alias{fig_clean}
\title{fig_clean
Removes color values close to white, also possible to remove text on plot, useful for gridlines on KM curves}
\usage{
fig_clean(fig.hsl, bg_lightness = 0.1, attempt_OCR = F, word_sensitivity = 30)
}
\arguments{
\item{fig.hsl}{array of pixels in HSL format (hue/saturation/lightness)}

\item{bg_lightness}{a lightness threshold value between 0 and 1; every pixel with lightness > bg_lightness is considered background and removed (default: 0.1).}

\item{attempt_OCR}{a logical indicating whether to attempt text recognition (default: False)}

\item{word_sensitivity}{A sensitivity percentage value for word recognition, potential pixel patterns are considered words if confidence > word_sensitivity (default: 30)}
}
\value{
fig.df: a dataframe with x,y and h,s,l values for each remaining pixel
}
\description{
fig_clean
Removes color values close to white, also possible to remove text on plot, useful for gridlines on KM curves
}
\examples{
# fig_clean(fig.hsl = figure, bg_lightness = 0.1, attempt_OCR = T, word_sensitivity = 30)
}
