% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_summarize.R
\name{fig_summarize}
\alias{fig_summarize}
\title{fig_summarize
Brings together the curve and range data from into a format that is readable by the survival function}
\usage{
fig_summarize(lines_vector, range_list)
}
\arguments{
\item{range_list}{: output from range_detect}

\item{lines_vector:}{output from lines_isolate}
}
\value{
a dataframe with id, times, St, curve
}
\description{
fig_summarize
Brings together the curve and range data from into a format that is readable by the survival function
}
\examples{
# fig_summarize(lines_vector, range_df)
}
