% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cleanplot.R
\name{fun_cleanplot}
\alias{fun_cleanplot}
\title{fun_cleanplot
Removes color values close to 0, also possible to remove text on plot}
\usage{
fun_cleanplot(fig.lst, i.sen, Wsen, OCR_words)
}
\arguments{
\item{fig.lst}{a list with figures from fun_idplot}

\item{i.sen}{how close a value is to one where we consider it a 0}

\item{Wsen}{Word sensitivity for image recognition}

\item{OCR_words}{a logical indicating wether to attempt text recognition}
}
\value{
fig.rgbclean: an array from removing pixels that are not relevant
}
\description{
fun_cleanplot
Removes color values close to 0, also possible to remove text on plot
}
\examples{
# fun_cleanplot(figlst, i.sen = 0.2, Wsen = 30, OCR_words = F)
}
