% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_range.R
\name{fun_range}
\alias{fun_range}
\title{fun_range
detects how X and y pixel values map to time and survival values respectively}
\usage{
fun_range(
  X_start,
  X_end,
  X_increment,
  Y_start,
  Y_increment,
  Y_end,
  step1_bw,
  step2_axis,
  Y_values_vertical
)
}
\arguments{
\item{X_start}{what does the x-axis start usually 0}

\item{X_end}{what does the x-axis end in}

\item{X_increment}{what does it go up by}

\item{Y_start}{what does the y-axis start}

\item{Y_increment}{what does it go up by}

\item{Y_end}{what does it end in}

\item{step1_bw}{from previous step}

\item{step2_axis}{from previous step}

\item{Y_values_vertical}{whether the y -axis labels are vertical or horizontal: does this make sense)}
}
\value{
a list with the Y_0pixel where the y-axis starts, y_increment, X_0pixel and X_increment.
}
\description{
fun_range
detects how X and y pixel values map to time and survival values respectively
}
