% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_readsurv.R
\name{fun_readsurv}
\alias{fun_readsurv}
\title{fun_readsurv
loads a jpeg or jpg or PNG file and turns into into arrays for further processing.}
\usage{
fun_readsurv(FilePath)
}
\arguments{
\item{FilePath}{character string indicating the location of the curve to be digitized}
}
\value{
fig.hsl array of pixels in HSL format (hue/saturation/lightness)
}
\description{
fun_readsurv
loads a jpeg or jpg or PNG file and turns into into arrays for further processing.
}
\examples{
 # fun_readsurv(FilePath = "~/curve1.jpeg")

}
