% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{fun_summary}
\alias{fun_summary}
\title{fun_summary
Brings together the data from get_final and fun_range into a format that is readable by the survival function}
\usage{
fun_summary(final_list, Step7_out)
}
\arguments{
\item{Step7_out}{: output from fun_range}

\item{final_list:}{output from get_final}
}
\value{
a dataframe with id, times, St, curve
}
\description{
fun_summary
Brings together the data from get_final and fun_range into a format that is readable by the survival function
}
\examples{
# fun_summary(get_out, range_df)
}
