% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tofinal.R
\name{get_tofinal}
\alias{get_tofinal}
\title{get_tofinal
removes points that don't fit the structure of a survvival curve based on e.gen}
\usage{
get_tofinal(res_list)
}
\arguments{
\item{res_list}{: output from overlap_detect}
}
\value{
res_vect: a dataframe that complies with the values of a survival curve
}
\description{
get_tofinal
removes points that don't fit the structure of a survvival curve based on e.gen
}
\examples{
# get_tofinal(res_list)
}
