% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_read.R
\name{img_read}
\alias{img_read}
\title{img_read
loads a jpeg or jpg or PNG file and turns into into arrays for further processing.}
\usage{
img_read(path)
}
\arguments{
\item{path}{character string indicating the location of the curve to be digitized}
}
\value{
fig.hsl array of pixels in HSL format (hue/saturation/lightness)
}
\description{
img_read
loads a jpeg or jpg or PNG file and turns into into arrays for further processing.
}
\examples{
 # img_read(path = "~/curve1.jpeg")

}
