% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_overlap.R
\name{impute_overlap}
\alias{impute_overlap}
\title{Impute Overlapping Values}
\usage{
impute_overlap(fig.curves = step4, size = 50)
}
\arguments{
\item{fig.curves}{A dataframe with x, y, and group values. Default value is 'step4'.}

\item{size}{An integer that specifies the size of number of time interval for imputation.. Default value is 50.}
}
\value{
A vector with a dataframe for each curve, stored in 'res_list'.
}
\description{
This function assists in resolving issues related to color curve overlaps that occur during the initial phases of a Kaplan-Meier (KM) study.
}
\examples{
# impute_overlap(fig.curves, size = 50)
}
