% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_censoring_removal.R
\name{line_censoring_removal}
\alias{line_censoring_removal}
\title{Remove Outliers from Color Curves}
\usage{
line_censoring_removal(dataframe)
}
\arguments{
\item{dataframe}{A dataframe with x, y, and group values. Default value is 'step4'.}
}
\value{
A dataframe for each curve, stored in 's4_processed'.
}
\description{
This function aids in the removal of outlier values within color curves, which can occur during the analysis of a Kaplan-Meier (KM) study. It works by censoring any y-values greater than the 90th percentile of the intra-group range for each group, thus ensuring a smoother and more interpretable curve.
}
\examples{
# line_censoring_removal(dataframe)
}
