% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_isolate.R
\name{lines_isolate}
\alias{lines_isolate}
\title{lines_isolate
Select the final singular y values for each curve (Note: this is where additional event detection could be performed)}
\usage{
lines_isolate(fig.curves)
}
\arguments{
\item{fig.curves}{: a dataframe with x, y and group values}
}
\value{
res_list: a vector with a dataframe for each curve
}
\description{
lines_isolate
Select the final singular y values for each curve (Note: this is where additional event detection could be performed)
}
\examples{
# lines_isolate(fig.curves)
}
