% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_detect.R
\name{overlap_detect}
\alias{overlap_detect}
\title{overlap_detect
this function detects potential overlap between curves}
\usage{
overlap_detect(step4)
}
\arguments{
\item{step4}{: output from fun_colordetection}
}
\value{
res_list: a dataframe with the detected values for all curves as well as potential overlap
}
\description{
overlap_detect
this function detects potential overlap between curves
}
\examples{
# overlap_detect(df_colordetect)
}
