% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_digitizer.R
\name{parent_digitizer}
\alias{parent_digitizer}
\title{Parent_digitizer a wrapper for all the digitizer functions}
\usage{
parent_digitizer(
  curve_location,
  sen,
  Wsen_i,
  OCR_words_i,
  num_curves1,
  x_start_i,
  x_end_i,
  x_incr,
  y_start_i,
  y_end_i,
  y_incr,
  Y_values_vertical
)
}
\arguments{
\item{curve_location}{: curve location}

\item{sen}{: a value between 0 and 1 indicating how close a value has to be to white to be considered part of the background}

\item{Wsen_i}{: word sensitivity for OCR word removal}

\item{OCR_words_i:}{logical indicating if needed to remove words}

\item{num_curves1:}{the number of curves in an image}

\item{x_start_i:}{lowest X-axis label value}

\item{x_end_i:}{highest X-axis label value}

\item{x_incr:}{what increments do the ticks go up in the x-axis}

\item{y_start_i:}{lowesst y-axis label}

\item{y_end_i:}{highrst y-axis lebel}

\item{y_incr:}{what increments do the ticks go up in the y-axis}

\item{Y_values_vertical:}{whether the y-axis labels are vertical}
}
\value{
a dataframe with id, times, St, curve
}
\description{
Parent_digitizer a wrapper for all the digitizer functions
}
\examples{
# out1<-parent_digitizer(curve_location = here::here("OS.jpg"), sen = 0.1, Wsen_i = 30 , OCR_words_i = F , num_curves1 = 2, x_start_i = 0, x_end_i = 10, x_incr = 1, y_start_i = 0 , y_end_i  = 1, y_incr = 0.2, Y_values_vertical = T)
}
