% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_detect.R
\name{range_detect}
\alias{range_detect}
\title{range_detect
detects how X and y pixel values map to time and survival values respectively}
\usage{
range_detect(
  step1_fig,
  step2_axes,
  x_start,
  x_end,
  x_increment,
  y_start,
  y_increment,
  y_end,
  y_text_vertical
)
}
\arguments{
\item{step1_fig}{figure array from previous step}

\item{step2_axes}{axes list from previous step}

\item{x_start}{what does the x-axis start usually 0}

\item{x_end}{what does the x-axis end in}

\item{x_increment}{what does it go up by (Note: increment between EVERY tick, including minor ones)}

\item{y_start}{what does the y-axis start}

\item{y_increment}{what does it go up by (Note: increment between EVERY tick, including minor ones)}

\item{y_end}{what does it end in}

\item{y_text_vertical}{whether the y -axis label text is vertical or horizontal)}
}
\value{
# a list with the Y_0pixel where the y-axis starts, y_increment, X_0pixel and x_increment.
}
\description{
range_detect
detects how X and y pixel values map to time and survival values respectively
}
\examples{
# range_detect(step1_fig, step2_axes, x_start, x_end, x_increment,y_start, y_increment, y_end, y_text_vertical)

}
