% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_digitize.R
\name{survival_digitize}
\alias{survival_digitize}
\title{survival_digitize a wrapper for the digitizer end-to-end process}
\usage{
survival_digitize(
  img_path,
  bg_lightness = 0.1,
  attempt_OCR = F,
  word_sensitivity = 30,
  num_curves,
  censoring,
  x_start,
  x_end,
  x_increment,
  y_start,
  y_end,
  y_increment,
  y_text_vertical
)
}
\arguments{
\item{img_path}{:  location of image file}

\item{bg_lightness}{: a lightness threshold value between 0 and 1; every pixel with lightness > bg_lightness is considered background and removed (default: 0.9)}

\item{Wsen:}{word sensitivity for OCR word removal}

\item{OCR_words:}{logical indicating if needed to remove words}

\item{num_curves:}{the number of curves in an image}

\item{censoring:}{locgical indicating if censoring is indicated by black marks}

\item{x_start:}{lowest X-axis label value}

\item{x_end:}{highest X-axis label value}

\item{x_increment:}{what increments do the ticks go up in the x-axis (Note: increment between EVERY tick, including minor ones)}

\item{y_start:}{lowest y-axis label}

\item{y_end:}{highest y-axis label}

\item{y_increment:}{what increments do the ticks go up in the y-axis (Note: increment between EVERY tick, including minor ones)}

\item{y_text_vertical:}{whether the y-axis labels are vertical (TRUE) or horizontal (FALSE)}
}
\value{
a dataframe with id, times, St, curve
}
\description{
survival_digitize a wrapper for the digitizer end-to-end process
}
\examples{
# survival_digitize(img_path = here::here("OS.jpg"), bg_lightness = 0.1, attempt_OCR = F , num_curves = 2, x_start = 0, x_end = 10, x_increment = 1, y_start = 0 , y_end  = 1, y_increment = 0.2, y_text_vertical = T)
}
