% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detect.R
\name{eventdetect}
\alias{eventdetect}
\title{eventdetect
detects potential events}
\usage{
eventdetect(res.df)
}
\arguments{
\item{res.df}{: a dataframe with x, y and group values}
}
\value{
res_list: a list with a dataframe for each curve
}
\description{
eventdetect
detects potential events
}
\examples{
# eventdetect(overlap_detectout)
}
