% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_summarize.R
\name{fig_summarize}
\alias{fig_summarize}
\title{Figure Summarization}
\usage{
fig_summarize(lines_vector, range_list, y_start = 0, y_end = 1)
}
\arguments{
\item{lines_vector}{The output from the `lines_isolate` function.}

\item{range_list}{The output from the `range_detect` function.}

\item{y_start}{Starting value of the y-axis, defaults to the global `y_start` variable.}

\item{y_end}{Ending value of the y-axis, defaults to the global `y_end` variable.}
}
\value{
A dataframe with columns: 'id', 'times', 'St', and 'curve'.
}
\description{
Integrates the curve and range data into a format that is readable by the survival function.
}
\examples{
# fig_summarize(lines_vector, range_list, y_start, y_end)
}
