% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cleanplot.R
\name{fun_cleanplot}
\alias{fun_cleanplot}
\title{fun_cleanplot
Removes color values close to white, also possible to remove text on plot, useful for gridlines on KM curves}
\usage{
fun_cleanplot(fig.hsl, i.sen, Wsen, OCR_words)
}
\arguments{
\item{fig.hsl}{array of pixels in HSL format (hue/saturation/lightness)}

\item{i.sen}{how close a value is to white to consider it part of the background}

\item{Wsen}{Word sensitivity for image recognition}

\item{OCR_words}{a logical indicating whether to attempt text recognition}
}
\value{
fig.df: a dataframe with x,y and h,s,l values for each remaining pixel
}
\description{
fun_cleanplot
Removes color values close to white, also possible to remove text on plot, useful for gridlines on KM curves
}
\examples{
# fun_cleanplot(figlst, i.sen = 0.2, Wsen = 30, OCR_words = F)
}
