% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_colordetect.R
\name{fun_colordetect}
\alias{fun_colordetect}
\title{fun_colordetect
detects the location of each curve in the image using medoids}
\usage{
fun_colordetect(fig.df, num_curves = NULL, black_marks = F)
}
\arguments{
\item{fig.df}{a dataframe with x,y and h,s,l values for each pixel}

\item{num_curves}{the number of curves that are on the figure to be digitized}

\item{black_marks}{logical indicating whether censoring occurs in a diff color}
}
\value{
fig.grp: a dataframe with x,y,h,s,l values and associated curve for each pixel
}
\description{
fun_colordetect
detects the location of each curve in the image using medoids
}
\examples{
# fun_colordetect(fig.df =  fig.df, num_curves = 3, black_marks = F)
}
