% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_colordetect.R
\name{fun_colordetect}
\alias{fun_colordetect}
\title{fun_colordetect
detects the location of each curve in the image using a kmeans algorithm}
\usage{
fun_colordetect(fig.list, num_curves)
}
\arguments{
\item{fig.list}{output from fun_cleanplot}

\item{num_curves}{the number of curves that are on the figure to be digitized}
}
\value{
comp1: a data frame with y,x,R,G,B color values and curve associated
}
\description{
fun_colordetect
detects the location of each curve in the image using a kmeans algorithm
}
\examples{
# fun_colordetect(fig.list =  fig.list, num_curves = 3)
}
