% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_detect.R
\name{overlap_detect}
\alias{overlap_detect}
\title{overlap_detect
this function detects potential overlap between curves and "fills in the gaps"}
\usage{
overlap_detect(fig.grp, k = 200)
}
\arguments{
\item{fig.grp:}{dataframe with x,y values and associated curve for each pixel}
}
\value{
res.df: a dataframe with the detected x,y, group values for all curves
}
\description{
overlap_detect
this function detects potential overlap between curves and "fills in the gaps"
}
\examples{
# overlap_detect(df_colordetect)
}
