% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_detect.R
\name{overlap_detect}
\alias{overlap_detect}
\title{overlap_detect
detects potential overlap between curves and "fills in the gaps" using knn similarity}
\usage{
overlap_detect(fig.grp, nr_neighbors = 20)
}
\arguments{
\item{fig.grp:}{dataframe with x,y values and associated curve for each pixel}

\item{nr_neighbors:}{how many nearby neighbors to consider when guessing the value of missing pixels (default: 100)}
}
\value{
res.df: a dataframe with the detected x,y, group values for all curves
}
\description{
overlap_detect
detects potential overlap between curves and "fills in the gaps" using knn similarity
}
\examples{
# overlap_detect(fig.dataframe, nr_neighbors = 20)
}
