% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_digitize.R
\name{survival_digitize}
\alias{survival_digitize}
\title{Survival Digitization}
\usage{
survival_digitize(
  img_path,
  bg_lightness = 0.3,
  attempt_OCR = F,
  word_sensitivity = 30,
  num_curves,
  censoring = T,
  x_start,
  x_end,
  x_increment,
  y_start,
  y_end,
  y_increment,
  y_text_vertical,
  nr_neighbors = 20,
  enhance = F,
  impute_size = 0
)
}
\arguments{
\item{img_path}{Location of the image file.}

\item{bg_lightness}{A lightness threshold value between 0 and 1. Every pixel with lightness > bg_lightness is considered background and removed. Default value is 0.3.}

\item{attempt_OCR}{Logical value indicating if Optical Character Recognition (OCR) should be attempted to remove words. Default value is FALSE.}

\item{word_sensitivity}{Sensitivity for OCR word removal. Default value is 30.}

\item{num_curves}{The number of curves in an image.}

\item{censoring}{Logical value indicating if censoring is represented by black marks. Default value is TRUE.}

\item{x_start}{Lowest X-axis label value.}

\item{x_end}{Highest X-axis label value.}

\item{x_increment}{Increment value for the X-axis ticks. Note: this should include increment between every tick, including minor ones.}

\item{y_start}{Lowest Y-axis label.}

\item{y_end}{Highest Y-axis label.}

\item{y_increment}{Increment value for the Y-axis ticks. Note: this should include increment between every tick, including minor ones.}

\item{y_text_vertical}{Boolean value indicating if the Y-axis labels are vertical (TRUE) or horizontal (FALSE).}

\item{nr_neighbors}{Number of neighbors in k-nearest neighbors (knn) when grouping pixels by colors. Default value is 50.}

\item{enhance}{Logical value indicating whether to convert HSL channels into the same scale. Default value is FALSE.}

\item{impute_size}{Size parameter for imputation.}
}
\value{
A dataframe with columns: 'id', 'times', 'St', and 'curve'.
}
\description{
A wrapper for the digitizer end-to-end process.
}
\examples{
# survival_digitize(img_path = here::here("OS.jpg"), bg_lightness = 0.1, attempt_OCR = FALSE, num_curves = 2, x_start = 0, x_end = 10, x_increment = 1, y_start = 0, y_end = 1, y_increment = 0.2, y_text_vertical = TRUE)
}
